switched to He weight initialization and fixed bug in Conv class constructor

+++ include/BaseLayer.hpp
@@ -57,11 +57,11 @@ class BaseLayer {
 		/* updates weights and biases using 
 		delta from backpropagation function */
 		void	update( const Matrix &prev_outputs,
-						double learning_rate,
+						float learning_rate,
 						int    timestep,
-						double l2_reg = 0.0,
-						double beta1 = 0.9,
-						double beta2 = 0.999 );
+						float l2_reg = 0.0,
+						float beta1 = 0.9,
+						float beta2 = 0.999 );
 
 		int 	getSize( void ) const;
 		void    setSize( int new_size );
